flasksecforge/templates/flask_api_boilerplate/requirements.txt
Flask>=2.3
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Cors
python-dotenv
marshmallow
marshmallow-sqlalchemy
gunicorn